feat: allow custom catalog pagination
