+++ packages/catalog/src/constants.ts
@@ -1,3 +1,8 @@
 // number of pages on either side of active
 export const PAGINATION_INNER_WINDOW = 3;
 export const PAGINATION_MAX_PAGES = PAGINATION_INNER_WINDOW * 2 + 1;
+export const DEFAULT_PAGE = 1;
+
+export const DEFAULT_PAGE_SIZE = 50;
+
+export const DEFAULT_HIDE_PAGE_LIST = false;
